+++ packages/botonic-plugin-flow-builder/src/action/ai-agent.ts
@@ -1,5 +1,14 @@
-import { FlowContent } from '../content-fields'
-import { FlowAiAgent } from '../content-fields/flow-ai-agent'
+import {
+  BotContext,
+  EventAction,
+  EventAiAgent,
+  InferenceResponse,
+} from '@botonic/core'
+
+import { FlowAiAgent, FlowContent } from '../content-fields'
+import { HtNodeWithContent } from '../content-fields/hubtype-fields'
+import { getFlowBuilderPlugin } from '../helpers'
+import { trackEvent } from '../tracking'
 import { GuardrailRule } from '../types'
 import { FlowBuilderContext } from './index'
 
@@ -46,6 +55,7 @@ export async function getContentsByAiAgent({
   if (!aiAgentResponse) {
     return []
   }
+  await trackAiAgentResponse(aiAgentResponse, request, aiAgentContent)
 
   if (aiAgentResponse.exit) {
     return []
@@ -54,4 +64,36 @@ export async function getContentsByAiAgent({
   aiAgentContent.responses = aiAgentResponse.messages
 
   return contents
+}
+
+async function trackAiAgentResponse(
+  aiAgentResponse: InferenceResponse,
+  request: BotContext,
+  aiAgentContent: FlowAiAgent
+) {
+  const flowBuilderPlugin = getFlowBuilderPlugin(request.plugins)
+  const flowId = flowBuilderPlugin.cmsApi.getNodeById<HtNodeWithContent>(
+    aiAgentContent.id
+  ).flow_id
+  const flowName = flowBuilderPlugin.getFlowName(flowId)
+
+  const event: EventAiAgent = {
+    action: EventAction.AiAgent,
+    flowThreadId: request.session.flow_thread_id!,
+    flowId: flowId,
+    flowName: flowName,
+    flowNodeId: aiAgentContent.id,
+    flowNodeContentId: aiAgentContent.name,
+    flowNodeIsMeaningful: true,
+    toolsExecuted: aiAgentResponse?.toolsExecuted ?? [],
+    memoryLength: aiAgentResponse?.memoryLength ?? 0,
+    inputMessageId: request.input.message_id!,
+    exit: aiAgentResponse?.exit ?? true,
+    inputGuardrailsTriggered: aiAgentResponse?.inputGuardrailsTriggered ?? [],
+    outputGuardrailsTriggered: [], //aiAgentResponse.outputGuardrailsTriggered,
+    error: aiAgentResponse.error,
+  }
+  const { action, ...eventArgs } = event
+
+  await trackEvent(request, action, eventArgs)
 }